% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{ipsum_examples}
\alias{ipsum_examples}
\title{Get a shinipsum example}
\usage{
ipsum_examples(which = NULL)
}
\arguments{
\item{which}{The example to run. If empty, all the available examples are listed.}
}
\value{
A path to the example.
}
\description{
Get a shinipsum example
}
\examples{
ipsum_examples()
}
