% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lorem_words}
\alias{lorem_words}
\title{Lorem test as vector}
\format{
An object of class \code{character} of length 13657.
}
\usage{
lorem_words
}
\description{
A long vector of words
}
\keyword{datasets}
