% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTable.R
\name{random_DT}
\alias{random_DT}
\title{A Random DT}
\usage{
random_DT(
  nrow,
  ncol,
  type = c("random", "numeric", "character", "numchar"),
  ...
)
}
\arguments{
\item{nrow}{number of row of the output}

\item{ncol}{number of cols of the output}

\item{type}{type of the columns, can be "random", "numeric", "character", "numchar". Default is random.}

\item{...}{arguments to be passed to DT::datatable}
}
\value{
a DT
}
\description{
This function creates a random DT::datatable, and can be passed into \code{renderDT} & \code{DTOutput}.
}
