% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphs.R
\name{random_dygraph}
\alias{random_dygraph}
\title{A Random Dygraph}
\usage{
random_dygraph(...)
}
\arguments{
\item{...}{args passed to \code{dygraph}}
}
\value{
a dygraph
}
\description{
This function returns a dygraph object, which can be passed to \code{renderDygraph} and \code{dygraphOutput}
}
