% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{random_ggplotly}
\alias{random_ggplotly}
\title{A Random ggplotly}
\usage{
random_ggplotly(...)
}
\arguments{
\item{...}{arg to pass to \code{random_ggplot}.}
}
\value{
a ggplotly
}
\description{
This function returns a ggplotly object, which can be passed to \code{renderPlotly} and \code{plotlyOutput}
}
