% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModel.R
\name{random_lm}
\alias{random_lm}
\title{A Random Linear Model}
\usage{
random_lm(nobs = 100, nx = 2)
}
\arguments{
\item{nobs}{Numeric. number of observation}

\item{nx}{Numeric. number of variables. Should be lower that nobs}
}
\value{
a model output
}
\description{
This function returns a model which can be passed to renderText or renderTable if pre-processed appropriately
}
