% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{random_print}
\alias{random_print}
\title{A Random print output}
\usage{
random_print(type = c("character", "numeric", "model", "table"))
}
\arguments{
\item{type}{type of the output ("character", "numeric", "model", "table")}
}
\value{
a random print
}
\description{
This function returns a random print output that can be passed to \code{renderPrint} and \code{verbatimTextOutput}.
}
