% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{random_table}
\alias{random_table}
\title{A Random Table}
\usage{
random_table(nrow, ncol, type = c("random", "numeric", "character", "numchar"))
}
\arguments{
\item{nrow}{number of row of the output}

\item{ncol}{number of cols of the output}

\item{type}{type of the columns, can be "random", "numeric", "character", "numchar". Default is random.}
}
\value{
a table
}
\description{
This function returns a table that can be passed to \code{renderTable} and \code{tableOutput}.
}
