% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinipsum-package.R
\docType{package}
\name{shinipsum-package}
\alias{shinipsum-package}
\alias{_PACKAGE}
\alias{shinipsum}
\title{Lorem-Ipsum-like Helpers for fast Shiny Prototyping}
\description{
The goal of shinipsum is to prototype your shiny apps quickly with these Lorem-Ipsum-like Helpers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/shinipsum}
  \item Report bugs at \url{https://github.com/Thinkr-open/shinipsum/issues}
}

}
\author{
Colin Fay \href{mailto:colin@thinkr.fr}{colin@thinkr.fr}
}
\keyword{internal}
