% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{random_ggplot}
\alias{random_ggplot}
\title{A Random ggplot}
\usage{
random_ggplot(
  type = c("random", "point", "bar", "boxplot", "col", "tile", "line", "bin2d",
    "contour", "density", "density_2d", "dotplot", "hex", "freqpoly", "histogram",
    "ribbon", "raster", "violin", "ts")
)
}
\arguments{
\item{type}{type of the geom. Can be any of "random", "point", "bar", "boxplot","col", "tile", "line", "bin2d", "contour", "density", "density_2d", "dotplot", "hex", "freqpoly", "histogram", "ribbon", "raster", "violin", "ts" (alias "timeseries") and defines the geom of the ggplot. Default is "random", and chooses a random geom for you. The "ts" type returns a time-series oriented plot, with a \code{Date} on the x axis.}
}
\value{
a ggplot
}
\description{
This function returns a ggplot object, which can be passed to \code{renderPlot} and \code{plotOutput}
}
