% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{random_image}
\alias{random_image}
\title{A Random Image}
\usage{
random_image()
}
\value{
an image
}
\description{
This function returns a random image that can be passed into \code{renderImage} and \code{plotOutput}.
}
