% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_image_ext.R
\name{random_image_ext}
\alias{random_image_ext}
\title{A Random External Image from the Lorem Picsum API}
\usage{
random_image_ext(width = 400, height = 400, seed = NULL)
}
\value{
an external image
}
\description{
This function returns an external random image from the Lorem Picsum API that can be used directly within the Shiny UI.
}
\examples{

random_image_ext(width = 400, height = 600, seed = "caramba")


}
