% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinipsum-package.R
\docType{package}
\name{shinipsum-package}
\alias{shinipsum}
\alias{shinipsum-package}
\title{shinipsum: Lorem-Ipsum-like Helpers for fast Shiny Prototyping}
\description{
Prototype your shiny apps quickly with these
    Lorem-Ipsum-like Helpers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/shinipsum}
  \item Report bugs at \url{https://github.com/Thinkr-open/shinipsum/issues}
}

}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (0000-0001-7343-1846)

Authors:
\itemize{
  \item Sebastien Rochette \email{sebastien@thinkr.fr} (0000-0002-1565-9313)
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder, funder]
}

}
\keyword{internal}
